+++ GISPython/xmlParamsHealper.py
@@ -1,21 +1,23 @@
 # -*- coding: utf-8 -*-
 """
-     Module for Json parameter file procedures
+     Module for xml parameter file procedures
 """
 
 import os
 import codecs
-import simplejson as json
+from lxml import etree
 from collections import OrderedDict
 
-class JsonParams(object):
-    """Json parameter reading support class"""
+utf8_parser = etree.XMLParser(encoding='utf-8')
+
+class XMLParams(object):
+    """xml parameter reading support class"""
     def __init__(self, Tool, ConfigFolder, ConfigFile):
         """Class initialization procedure
 
         Args:
             self: The reserved object 'self'
-            Tool: GEOPython tool (optional)
+            Tool: GISPythonTool (Optional)
             ConfigFolder: Configuration file storing directory
             ConfigFile: Name of the configuration file (without extension)
         """
@@ -29,8 +31,6 @@ class JsonParams(object):
                 self.ConfigPath = os.path.join(ConfigFolder, ConfigFile)
             else:
                 self.ConfigPath = ConfigFile
-        if not self.ConfigPath.lower()[-4:] == "json":
-            self.ConfigPath = self.ConfigPath + '.json'
         self.Params = []
 
     def GetParams(self):
@@ -39,11 +39,10 @@ class JsonParams(object):
         Args:
             self: The reserved object 'self'
         """
-        f = codecs.open(self.ConfigPath, 'r', 'utf-8')
-        JsonString = f.read()
-        f.close()
-        J = json.loads(JsonString, object_pairs_hook=OrderedDict)
-        self.Params = J
+        with open(self.ConfigPath, 'r') as xmlfile:
+            xmllines = xmlfile.readlines()
+        doc = etree.ElementTree(etree.XML("".join(xmllines)))
+        self.Params = doc
         return self.Params
 
     def WriteParams(self):
@@ -52,29 +51,25 @@ class JsonParams(object):
         Args:
             self: The reserved object 'self'
         """
-        JsonString = json.dumps(self.Params, sort_keys=True, indent=4 * ' ')
-        f = codecs.open(self.ConfigPath, 'w', 'utf-8')
-        f.write(JsonString)
-        f.close()
+        with open(self.ConfigPath, 'w') as xmlfile:
+            # xmlString = etree.tostring(self.Params, pretty_print=True, method="xml", xml_declaration=True, encoding="utf-8")
+            self.Params.write(xmlfile, pretty_print=True, method="xml", xml_declaration=True, encoding="utf-8")
 
-    def UpdateValueByPath(self, path, Value, valueIsStringJson = False):
+
+    def UpdateValueByPath(self, path, Value, index = 0):
         elem = self.Params
+        elem.xpath(path)[index].text = Value
 
-        if valueIsStringJson:
-            Value = json.loads(Value, object_pairs_hook=OrderedDict)
-
-        pathList = path.strip("\\").split("\\")
-        for x in pathList[:-1]:
-                elem = elem[x]
-        LastKey = pathList[-1:][0]
-
-        elem[LastKey] = Value
-
+    def UpdateAtributeByPath(self, path, atribute, Value, index = 0):
+        elem = self.Params
+        elem.xpath(path)[index].attrib[atribute] = Value
 
     def GetValueByPath(self, path):
         elem = self.Params
-        pathList = path.strip("\\").split("\\")
-        for x in pathList:
-                elem = elem[x]
+        elem = elem.xpath(path)[0].text
+        return elem
 
+    def GetAtributeByPath(self, path, atribute):
+        elem = self.Params
+        elem = elem.xpath(path)[0].attrib[atribute]
         return elem